[feat] 모임 탈퇴 로직 테스트코드 작성
